Extract RepoConfigParser from Templater
Currently performs the parsing of the loaded base config for a node.
